// models/command.ts
export enum Command {
    SELECT = 'SELECT',
    UPDATE = 'UPDATE',
    INSERT_INTO = 'INSERT INTO',
    CREATE_DATABASE = 'CREATE DATABASE',
    ALTER_DATABASE = 'ALTER DATABASE',
    CREATE_TABLE = 'CREATE TABLE',
    ALTER_TABLE = 'ALTER TABLE',
    DROP_TABLE = 'DROP TABLE',
    CREATE_INDEX = 'CREATE INDEX',
    DROP_INDEX = 'DROP INDEX',
}